Data-crypto-main/Data-crypto-main/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
psycopg2-binary>=2.9.7
redis>=5.0.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
aiohttp>=3.8.0
tenacity>=8.2.0
web3>=6.0.0
asyncpg>=0.28.0
python-dotenv>=1.0.0
polars>=0.20.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.1
